Updated graph to use useContract hooks.  Merged most of Papi's cli to a hook for the liquidation contract.  Added in initial panels for placing bids and shwoing current bids.
