+++ package.json
@@ -37,6 +37,7 @@
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
     "@mui/material": "^5.3.0",
+    "@mui/x-data-grid": "^5.3.0",
     "@terra-money/terra.js": "^3.0.2",
     "@terra-money/wallet-provider": "^3.6.5",
     "bignumber.js": "^9.0.2",
