+++ src/index.tsx
@@ -4,6 +4,25 @@ import { BrowserRouter } from "react-router-dom";
 import ReactDOM from 'react-dom';
 import './style.css'
 
+import {
+    Chart as ChartJS,
+    CategoryScale,
+    LinearScale,
+    BarElement,
+    Title,
+    Tooltip,
+    Legend
+} from 'chart.js';
+
+ChartJS.register(
+    CategoryScale,
+    LinearScale,
+    BarElement,
+    Title,
+    Tooltip,
+    Legend
+);
+
 getChainOptions().then((chainOptions) => {
   ReactDOM.render(
     <BrowserRouter>
